# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for bmstu dormitory app.",
        "title": "BMSTU-WEB API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
...
        "/api/v1/students/": {
            "get": {
                "description": "View full information about students have lived in dormitory.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all students in dormitory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.StudentResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new student in user and student base.",
                "produces": [
                    "application/json"
                ],
                "summary": "Add new student to base",
                "parameters": [
                    {
                        "description": "student base information",
                        "name": "user-params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddNewStudentRequestMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.Student"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Пользователь с таким логином уже существует!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/students/{stud-number}/": {
            "put": {
                "description": "Change in database group information about student.",
                "produces": [
                    "application/json"
                ],
                "summary": "Change student group for student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Number",
                        "name": "stud-number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New student group",
                        "name": "new-group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeStudentGroupRequestMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные о студенте успешно обновлены!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Студент не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/students/{stud-number}/rooms/": {
            "post": {
                "description": "Settle student in certain room.",
                "produces": [
                    "application/json"
                ],
                "summary": "Settle student in room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Number",
                        "name": "stud-number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New student room ID",
                        "name": "room-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SettleInRoomRequestMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные о студенте успешно обновлены!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Студент не найден\" | \"Комната не найдена",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Студент уже живёт в другой комнате!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Settle student in certain room.",
                "produces": [
                    "application/json"
                ],
                "summary": "Evic student from current room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Number",
                        "name": "stud-number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные о студенте успешно обновлены!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Студент не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Студент уже нигде не живёт!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/students/{stud-number}/things/{mark-number}/": {
            "post": {
                "description": "Give thing to student without changing its location.",
                "produces": [
                    "application/json"
                ],
                "summary": "Give thing to student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Number",
                        "name": "stud-number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Mark number of thing",
                        "name": "mark-number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вещь передана!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Студент не найден\" | \"Вещь не найдена!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Вещь уже у другого студента!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Return thing from student without changing thing location.",
                "summary": "Return thing from student.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Number",
                        "name": "stud-number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Mark number of thing",
                        "name": "mark-number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вещь передана!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Студент не найден\" | \"Вещь не найдена!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Вещь и так была не у студента!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddNewStudentRequestMessage": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
        "models.ChangeStudentGroupRequestMessage": {
            "type": "object",
            "properties": {
                "newGroup": {
                    "type": "string"
                }
            }
        },
        "models.SettleInRoomRequestMessage": {
            "type": "object",
            "properties": {
                "roomID": {
                    "type": "integer"
                }
            }
        },
        "models.ShortResponseMessage": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
...
        "objects.Student": {
            "type": "object",
            "properties": {
                "accID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roomID": {
                    "type": "integer"
                },
                "studentGroup": {
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "objects.StudentResponseDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "studentGroup": {
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
    }
}